Initial commit:
Unfortunately, Prof. Matti Karjalainen, the author of this project
passed away in 2010.  I received permission to republish this source
code under a GPL license.  Matti made two releases.

"I'm sure that Matti would like the rest of the world to be able to
use his BC ideas. Please go ahead and continue its development
and distribution, if that's of interest to you.
Nobody at the university would continue doing it, I'm afraid."

This is a copy of Matti's release from Feb 15, 2008.
(designed for Windows-CLISP).

Signed-off-by: Damien Zammit <damien@zamaudio.com>