Release June 30 2008.
Unfortunately, Prof. Matti Karjalainen, the author of this project
passed away in 2010.  I received permission to republish this source
code under a GPL license.  Matti made two releases.

"I'm sure that Matti would like the rest of the world to be able to
use his BC ideas. Please go ahead and continue its development
and distribution, if that's of interest to you.
Nobody at the university would continue doing it, I'm afraid."

This is an exact copy of Matti's second release from June 30, 2008.
(designed for LispWorks Lisp for MacOSX, Windows and GNU/Linux).

  Generic distribution with GUI support, including examples of GUI controls
  Tested on Mac OSX 10.4&10.5 (Universal), Windows (XP). On Linux not tested yet
  For a realtime guitar synthesizer, see BC3/Applications/guitar.txt

Signed-off-by: Damien Zammit <damien@zamaudio.com>